// src/types/entities.d.ts
interface BaseProfile {
  readonly id: string;
  readonly businessId: number;
  readonly locale: string;
  readonly siteDomain: string;
  readonly siteId: number;
  readonly siteInternalHostname: string;
  readonly uid: number;
  readonly meta: {
    readonly entityType: {
      readonly id: string;
      readonly uid: number;
    };
    readonly locale: string;
  };
  readonly _site: SiteProfile;
}

export interface SiteProfile extends BaseProfile {
  readonly name: string;
  readonly c_brand?: string;
  readonly c_copyrightMessage?: string;
  readonly c_facebook?: string;
  readonly c_instagram?: string;
  readonly c_youtube?: string;
  readonly c_twitter?: string;
  readonly c_linkedIn?: string;
  readonly c_footerLinks?: CTA[];
  readonly c_cFPLogic?: boolean;
  readonly c_cFPCopyRightText?: RTF2;
  readonly c_header?: {
    readonly logo?: Image;
    readonly logoLink?: string;
    readonly links?: CTA[];
  };
  readonly c_nearbySectionAPIKey?: string;
  readonly c_searchExperienceAPIKey?: string;
  readonly c_reviewsAPIKey?: string;
  readonly c_searchPage?: {
    readonly slug?: string;
  };
  readonly c_bannerText?: string;
  readonly c_bannerURL?: string;
  readonly c_placeholderPhoto?: Image;
  readonly c_footerColumns?: FooterColumn[];
  readonly logo?: ComplexImage;
  readonly c_copyrightText?: RTF2;
  readonly c_companyPhoneNumber: CTA;
  readonly c_footerPhoto: ComplexImage;
  readonly c_defaultResultCard: ResultCard;
}

export interface FooterColumn {
  readonly title?: string;
  readonly footerLink?: CTA[];
  readonly footerText?: string;
}

export interface RTF2 {
  readonly json?: object;
}
// ...
export interface ResultCard {
  readonly photo?: Image;
  readonly name?: string;
  readonly title: string;
  readonly phone?: CTA;
  readonly cTA1?: CTA;
  readonly cTA2?: CTA;
}
// ...
export type TemplateProps<T = Record<string, unknown>> = Omit<
  InternalTemplateProps,
  "document"
> & {
  document: T;
};
export type TemplateRenderProps<T = Record<string, unknown>> = Omit<
  InternalTemplateRenderProps,
  "document"
> &
  TemplateProps<T> & {
    translations?: Resource;
  };